Claude 3/messages/vision integration for anthropic
