+++ libs/community/pyproject.toml
@@ -70,7 +70,7 @@ sqlite-vss = {version = "^0.1.2", optional = true}
 motor = {version = "^3.3.1", optional = true}
 timescale-vector = {version = "^0.0.1", optional = true}
 typer = {version= "^0.9.0", optional = true}
-anthropic = {version = "^0.3.11", optional = true}
+anthropic = {version = "^0.16.0", optional = true}
 aiosqlite = {version = "^0.19.0", optional = true}
 rspace_client = {version = "^2.5.0", optional = true}
 upstash-redis = {version = "^0.15.0", optional = true}
@@ -152,7 +152,7 @@ openai = "^1"
 python-dotenv = "^1.0.0"
 cassio = "^0.1.0"
 tiktoken = ">=0.3.2,<0.6.0"
-anthropic = "^0.3.11"
+anthropic = "^0.16.0"
 langchain-core = { path = "../core", develop = true }
 fireworks-ai = "^0.9.0"
 
@@ -267,7 +267,7 @@ extended_testing = [
  "rdflib",
  "tidb-vector",
  "cloudpickle",
- "friendli-client"
+ "friendli-client",
 ]
 
 [tool.ruff]
