+++ libs/community/tests/unit_tests/chat_models/test_anthropic.py
@@ -1,14 +1,9 @@
 """Test Anthropic Chat API wrapper."""
 import os
-from typing import List
 
 import pytest
-from langchain_core.messages import AIMessage, BaseMessage, HumanMessage, SystemMessage
 
 from langchain_community.chat_models import ChatAnthropic
-from langchain_community.chat_models.anthropic import (
-    convert_messages_to_prompt_anthropic,
-)
 
 os.environ["ANTHROPIC_API_KEY"] = "foo"
 
@@ -50,26 +45,3 @@ def test_anthropic_initialization() -> None:
     # Verify that chat anthropic can be initialized using a secret key provided
     # as a parameter rather than an environment variable.
     ChatAnthropic(model="test", anthropic_api_key="test")
-
-
-@pytest.mark.parametrize(
-    ("messages", "expected"),
-    [
-        ([HumanMessage(content="Hello")], "\n\nHuman: Hello\n\nAssistant:"),
-        (
-            [HumanMessage(content="Hello"), AIMessage(content="Answer:")],
-            "\n\nHuman: Hello\n\nAssistant: Answer:",
-        ),
-        (
-            [
-                SystemMessage(content="You're an assistant"),
-                HumanMessage(content="Hello"),
-                AIMessage(content="Answer:"),
-            ],
-            "You're an assistant\n\nHuman: Hello\n\nAssistant: Answer:",
-        ),
-    ],
-)
-def test_formatting(messages: List[BaseMessage], expected: str) -> None:
-    result = convert_messages_to_prompt_anthropic(messages)
-    assert result == expected
